refactor(shared): Fase 1b — +7 globais universais -> rts-shared
regexp json json5 intl dom_exception global_this date(global). Limpos: so
engine heap + ns shared-sibling (regexp->regex, date->date, ambos ja em shared)
+ json/json5 metadata (externs __RTS_FN_NS_JSON_* resolvem em link-time).

- git mv globals/{7} -> rts-shared/src/globals/
- gc::handles -> rts_engine::heap::handles
- crate::namespaces::regex -> crate::regex; crate::namespaces::date -> crate::date
- rts-shared Cargo += indexmap + time (local-offset)
- facade: namespaces/globals/mod.rs pub use rts_shared::globals::<g>

rts-shared agora: 13 ns + 15 globais. Restam globais no runtime: cluster
(function/collections/proxy/error/string + reflect) + platform-divergent
(console/timers/fetch/performance/blob/headers/... -> step D rts-std).

Gate: cargo build --release OK (sem warnings); rts.exe test 1710/1710.
