+++ crates/rts-runtime/src/namespaces/globals/mod.rs
@@ -6,28 +6,28 @@ pub mod abort;
 pub use rts_shared::globals::bigint;
 pub mod blob;
 pub mod dataview;
-pub mod dom_exception;
+pub use rts_shared::globals::dom_exception;
 pub mod event_target;
 pub mod form_data;
 pub use rts_shared::globals::boolean;
 pub mod console;
-pub mod date;
+pub use rts_shared::globals::date;
 pub use rts_shared::globals::number;
 pub mod error;
 pub mod events;
 pub mod fetch;
 pub mod function;
-pub mod global_this;
+pub use rts_shared::globals::global_this;
 pub mod headers;
-pub mod intl;
-pub mod json;
-pub mod json5;
+pub use rts_shared::globals::intl;
+pub use rts_shared::globals::json;
+pub use rts_shared::globals::json5;
 pub mod message_channel;
 pub mod performance;
 pub mod proxy;
 pub mod readable_stream;
 pub mod reflect;
-pub mod regexp;
+pub use rts_shared::globals::regexp;
 pub mod string;
 pub use rts_shared::globals::symbol;
 pub mod text_encoding;
