+++ crates/rts-shared/Cargo.toml
@@ -13,3 +13,5 @@ edition = "2024"
 rts-engine = { path = "../rts-engine" }
 regex = "1"
 fancy-regex = "0.13"
+indexmap = "2.14.0"
+time = { version = "0.3", features = ["local-offset"] }
