+++ crates/rts-shared/src/globals/date/instance.rs
@@ -3,7 +3,7 @@
 //! Each `Date` instance is stored as `Entry::DateMs(i64)` in the HandleTable,
 //! where the i64 is milliseconds since Unix epoch (UTC).
 
-use crate::namespaces::gc::handles::{alloc_entry, with_entry, Entry};
+use rts_engine::heap::handles::{alloc_entry, with_entry, Entry};
 
 // ── Helper ────────────────────────────────────────────────────────────────────
 
@@ -19,7 +19,7 @@ fn get_ms(handle: u64) -> i64 {
 /// `new Date()` — current Unix timestamp in milliseconds.
 #[unsafe(no_mangle)]
 pub extern "C" fn __RTS_FN_GL_DATE_NEW_NOW() -> u64 {
-    let ms = crate::namespaces::date::__RTS_FN_NS_DATE_NOW_MS();
+    let ms = crate::date::__RTS_FN_NS_DATE_NOW_MS();
     alloc_entry(Entry::DateMs(ms))
 }
 
@@ -32,7 +32,7 @@ pub extern "C" fn __RTS_FN_GL_DATE_NEW_FROM_MS(ms: i64) -> u64 {
 /// `new Date(iso_str)` — from ISO 8601 string (`ptr`/`len` ABI).
 #[unsafe(no_mangle)]
 pub extern "C" fn __RTS_FN_GL_DATE_NEW_FROM_ISO(ptr: i64, len: i64) -> u64 {
-    let ms = crate::namespaces::date::__RTS_FN_NS_DATE_FROM_ISO(ptr as *const u8, len);
+    let ms = crate::date::__RTS_FN_NS_DATE_FROM_ISO(ptr as *const u8, len);
     alloc_entry(Entry::DateMs(ms))
 }
 
@@ -57,7 +57,7 @@ pub extern "C" fn __RTS_FN_GL_DATE_NEW_FROM_FIELDS(
     let s = if second.is_nan() { 0 } else { second as i64 };
     let frac_ms = if ms.is_nan() { 0 } else { ms as i64 };
     let total_ms_pretend_utc =
-        crate::namespaces::date::__RTS_FN_NS_DATE_FROM_PARTS(y, mo, d, h, mi, s, frac_ms);
+        crate::date::__RTS_FN_NS_DATE_FROM_PARTS(y, mo, d, h, mi, s, frac_ms);
     // (cross-runtime #172) `new Date(year, mon, day, ...)` em JS spec
     // interpreta os componentes como LOCAL time. pack() retorna ms como
     // se fosse UTC, entao convertemos: utc_ms = local_components_ms -
@@ -82,47 +82,47 @@ pub extern "C" fn __RTS_FN_GL_DATE_VALUE_OF(handle: u64) -> i64 {
 
 #[unsafe(no_mangle)]
 pub extern "C" fn __RTS_FN_GL_DATE_GET_FULL_YEAR(handle: u64) -> i64 {
-    crate::namespaces::date::__RTS_FN_NS_DATE_YEAR(get_ms(handle))
+    crate::date::__RTS_FN_NS_DATE_YEAR(get_ms(handle))
 }
 
 #[unsafe(no_mangle)]
 pub extern "C" fn __RTS_FN_GL_DATE_GET_MONTH(handle: u64) -> i64 {
-    crate::namespaces::date::__RTS_FN_NS_DATE_MONTH(get_ms(handle))
+    crate::date::__RTS_FN_NS_DATE_MONTH(get_ms(handle))
 }
 
 #[unsafe(no_mangle)]
 pub extern "C" fn __RTS_FN_GL_DATE_GET_DATE(handle: u64) -> i64 {
-    crate::namespaces::date::__RTS_FN_NS_DATE_DAY(get_ms(handle))
+    crate::date::__RTS_FN_NS_DATE_DAY(get_ms(handle))
 }
 
 #[unsafe(no_mangle)]
 pub extern "C" fn __RTS_FN_GL_DATE_GET_HOURS(handle: u64) -> i64 {
-    crate::namespaces::date::__RTS_FN_NS_DATE_HOUR(get_ms(handle))
+    crate::date::__RTS_FN_NS_DATE_HOUR(get_ms(handle))
 }
 
 #[unsafe(no_mangle)]
 pub extern "C" fn __RTS_FN_GL_DATE_GET_MINUTES(handle: u64) -> i64 {
-    crate::namespaces::date::__RTS_FN_NS_DATE_MINUTE(get_ms(handle))
+    crate::date::__RTS_FN_NS_DATE_MINUTE(get_ms(handle))
 }
 
 #[unsafe(no_mangle)]
 pub extern "C" fn __RTS_FN_GL_DATE_GET_SECONDS(handle: u64) -> i64 {
-    crate::namespaces::date::__RTS_FN_NS_DATE_SECOND(get_ms(handle))
+    crate::date::__RTS_FN_NS_DATE_SECOND(get_ms(handle))
 }
 
 #[unsafe(no_mangle)]
 pub extern "C" fn __RTS_FN_GL_DATE_GET_MILLISECONDS(handle: u64) -> i64 {
-    crate::namespaces::date::__RTS_FN_NS_DATE_MILLISECOND(get_ms(handle))
+    crate::date::__RTS_FN_NS_DATE_MILLISECOND(get_ms(handle))
 }
 
 #[unsafe(no_mangle)]
 pub extern "C" fn __RTS_FN_GL_DATE_GET_DAY(handle: u64) -> i64 {
-    crate::namespaces::date::__RTS_FN_NS_DATE_WEEKDAY(get_ms(handle))
+    crate::date::__RTS_FN_NS_DATE_WEEKDAY(get_ms(handle))
 }
 
 #[unsafe(no_mangle)]
 pub extern "C" fn __RTS_FN_GL_DATE_TO_ISO_STRING(handle: u64) -> u64 {
-    crate::namespaces::date::__RTS_FN_NS_DATE_TO_ISO(get_ms(handle))
+    crate::date::__RTS_FN_NS_DATE_TO_ISO(get_ms(handle))
 }
 
 /// `Date.prototype.toString()` — formato JS spec:
@@ -130,7 +130,7 @@ pub extern "C" fn __RTS_FN_GL_DATE_TO_ISO_STRING(handle: u64) -> u64 {
 /// RTS sempre UTC, entao tz fixo.
 #[unsafe(no_mangle)]
 pub extern "C" fn __RTS_FN_GL_DATE_TO_STRING(handle: u64) -> u64 {
-    use crate::namespaces::date::*;
+    use crate::date::*;
     let ms = with_entry(handle, |e| match e {
         Some(Entry::DateMs(v)) => *v,
         _ => 0,
@@ -169,7 +169,7 @@ pub extern "C" fn __RTS_FN_GL_DATE_TO_LOCALE_DATE_STRING(handle: u64) -> u64 {
     // formato mais comum em vez do ISO completo. JS spec deixa em aberto
     // o formato exato — qualquer impl deve gerar string parseable e
     // estavel pra o user.
-    use crate::namespaces::date::*;
+    use crate::date::*;
     let ms = with_entry(handle, |e| match e {
         Some(Entry::DateMs(v)) => *v,
         _ => 0,
@@ -248,7 +248,7 @@ pub(crate) fn local_offset_seconds_at(_ms: i64) -> i32 {
 /// (#552) `toUTCString()` — formato RFC 1123: "Day, DD Mon YYYY HH:MM:SS GMT".
 #[unsafe(no_mangle)]
 pub extern "C" fn __RTS_FN_GL_DATE_TO_UTC_STRING(handle: u64) -> u64 {
-    use crate::namespaces::date::*;
+    use crate::date::*;
     let ms = with_entry(handle, |e| match e {
         Some(Entry::DateMs(v)) => *v,
         _ => 0,
@@ -283,7 +283,7 @@ pub extern "C" fn __RTS_FN_GL_DATE_TO_UTC_STRING(handle: u64) -> u64 {
 /// que usam o formato `<weekday> <month> <day> <year>` definido pela spec.
 #[unsafe(no_mangle)]
 pub extern "C" fn __RTS_FN_GL_DATE_TO_DATE_STRING(handle: u64) -> u64 {
-    use crate::namespaces::date::*;
+    use crate::date::*;
     let ms = with_entry(handle, |e| match e {
         Some(Entry::DateMs(v)) => *v,
         _ => 0,
@@ -310,10 +310,10 @@ pub extern "C" fn __RTS_FN_GL_DATE_TO_DATE_STRING(handle: u64) -> u64 {
 // Cada setter le os parts atuais, substitui o slot, recompoe ms,
 // escreve de volta no slot. Retorna novos ms (JS spec).
 
-use crate::namespaces::gc::handles::with_entry_mut;
+use rts_engine::heap::handles::with_entry_mut;
 
 fn ms_to_parts(ms: i64) -> (i64, i64, i64, i64, i64, i64, i64) {
-    use crate::namespaces::date::*;
+    use crate::date::*;
     (
         __RTS_FN_NS_DATE_YEAR(ms),
         __RTS_FN_NS_DATE_MONTH(ms),
@@ -335,7 +335,7 @@ fn set_ms(handle: u64, new_ms: i64) {
 
 #[unsafe(no_mangle)]
 pub extern "C" fn __RTS_FN_GL_DATE_SET_FULL_YEAR(handle: u64, year: i64) -> i64 {
-    use crate::namespaces::date::__RTS_FN_NS_DATE_FROM_PARTS;
+    use crate::date::__RTS_FN_NS_DATE_FROM_PARTS;
     let (_, mo, d, h, mi, s, ms) = ms_to_parts(get_ms(handle));
     let new_ms = __RTS_FN_NS_DATE_FROM_PARTS(year, mo, d, h, mi, s, ms);
     set_ms(handle, new_ms);
@@ -344,7 +344,7 @@ pub extern "C" fn __RTS_FN_GL_DATE_SET_FULL_YEAR(handle: u64, year: i64) -> i64 
 
 #[unsafe(no_mangle)]
 pub extern "C" fn __RTS_FN_GL_DATE_SET_MONTH(handle: u64, month: i64) -> i64 {
-    use crate::namespaces::date::__RTS_FN_NS_DATE_FROM_PARTS;
+    use crate::date::__RTS_FN_NS_DATE_FROM_PARTS;
     let (y, _, d, h, mi, s, ms) = ms_to_parts(get_ms(handle));
     let new_ms = __RTS_FN_NS_DATE_FROM_PARTS(y, month, d, h, mi, s, ms);
     set_ms(handle, new_ms);
@@ -353,7 +353,7 @@ pub extern "C" fn __RTS_FN_GL_DATE_SET_MONTH(handle: u64, month: i64) -> i64 {
 
 #[unsafe(no_mangle)]
 pub extern "C" fn __RTS_FN_GL_DATE_SET_DATE(handle: u64, day: i64) -> i64 {
-    use crate::namespaces::date::__RTS_FN_NS_DATE_FROM_PARTS;
+    use crate::date::__RTS_FN_NS_DATE_FROM_PARTS;
     let (y, mo, _, h, mi, s, ms) = ms_to_parts(get_ms(handle));
     let new_ms = __RTS_FN_NS_DATE_FROM_PARTS(y, mo, day, h, mi, s, ms);
     set_ms(handle, new_ms);
@@ -362,7 +362,7 @@ pub extern "C" fn __RTS_FN_GL_DATE_SET_DATE(handle: u64, day: i64) -> i64 {
 
 #[unsafe(no_mangle)]
 pub extern "C" fn __RTS_FN_GL_DATE_SET_HOURS(handle: u64, hour: i64) -> i64 {
-    use crate::namespaces::date::__RTS_FN_NS_DATE_FROM_PARTS;
+    use crate::date::__RTS_FN_NS_DATE_FROM_PARTS;
     let (y, mo, d, _, mi, s, ms) = ms_to_parts(get_ms(handle));
     let new_ms = __RTS_FN_NS_DATE_FROM_PARTS(y, mo, d, hour, mi, s, ms);
     set_ms(handle, new_ms);
@@ -371,7 +371,7 @@ pub extern "C" fn __RTS_FN_GL_DATE_SET_HOURS(handle: u64, hour: i64) -> i64 {
 
 #[unsafe(no_mangle)]
 pub extern "C" fn __RTS_FN_GL_DATE_SET_MINUTES(handle: u64, min: i64) -> i64 {
-    use crate::namespaces::date::__RTS_FN_NS_DATE_FROM_PARTS;
+    use crate::date::__RTS_FN_NS_DATE_FROM_PARTS;
     let (y, mo, d, h, _, s, ms) = ms_to_parts(get_ms(handle));
     let new_ms = __RTS_FN_NS_DATE_FROM_PARTS(y, mo, d, h, min, s, ms);
     set_ms(handle, new_ms);
@@ -380,7 +380,7 @@ pub extern "C" fn __RTS_FN_GL_DATE_SET_MINUTES(handle: u64, min: i64) -> i64 {
 
 #[unsafe(no_mangle)]
 pub extern "C" fn __RTS_FN_GL_DATE_SET_SECONDS(handle: u64, sec: i64) -> i64 {
-    use crate::namespaces::date::__RTS_FN_NS_DATE_FROM_PARTS;
+    use crate::date::__RTS_FN_NS_DATE_FROM_PARTS;
     let (y, mo, d, h, mi, _, ms) = ms_to_parts(get_ms(handle));
     let new_ms = __RTS_FN_NS_DATE_FROM_PARTS(y, mo, d, h, mi, sec, ms);
     set_ms(handle, new_ms);
@@ -389,7 +389,7 @@ pub extern "C" fn __RTS_FN_GL_DATE_SET_SECONDS(handle: u64, sec: i64) -> i64 {
 
 #[unsafe(no_mangle)]
 pub extern "C" fn __RTS_FN_GL_DATE_SET_MILLISECONDS(handle: u64, ms_in: i64) -> i64 {
-    use crate::namespaces::date::__RTS_FN_NS_DATE_FROM_PARTS;
+    use crate::date::__RTS_FN_NS_DATE_FROM_PARTS;
     let (y, mo, d, h, mi, s, _) = ms_to_parts(get_ms(handle));
     let new_ms = __RTS_FN_NS_DATE_FROM_PARTS(y, mo, d, h, mi, s, ms_in);
     set_ms(handle, new_ms);
@@ -414,7 +414,7 @@ pub extern "C" fn __RTS_FN_GL_DATE_TO_JSON(handle: u64) -> u64 {
 /// retorna o formato mais comum em vez do ISO completo.
 #[unsafe(no_mangle)]
 pub extern "C" fn __RTS_FN_GL_DATE_TO_LOCALE_STRING(handle: u64) -> u64 {
-    use crate::namespaces::date::*;
+    use crate::date::*;
     let ms = with_entry(handle, |e| match e {
         Some(Entry::DateMs(v)) => *v,
         _ => 0,
@@ -436,7 +436,7 @@ pub extern "C" fn __RTS_FN_GL_DATE_TO_LOCALE_STRING(handle: u64) -> u64 {
 #[unsafe(no_mangle)]
 pub extern "C" fn __RTS_FN_GL_DATE_TO_LOCALE_TIME_STRING(handle: u64) -> u64 {
     // Locale time string: `HH:MM:SS` (sem ms nem timezone).
-    use crate::namespaces::date::*;
+    use crate::date::*;
     let ms = with_entry(handle, |e| match e {
         Some(Entry::DateMs(v)) => *v,
         _ => 0,
