+++ crates/rts-shared/src/globals/dom_exception/mod.rs
@@ -12,7 +12,7 @@ use indexmap::IndexMap;
 use rts_engine::abi::ty::{Handle, I64};
 use rts_engine::{AbiType, Engine, FnPtr, Member, MemberFlags, MemberKind, Sig};
 
-use crate::namespaces::gc::handles::{alloc_entry, with_entry, Entry};
+use rts_engine::heap::handles::{alloc_entry, with_entry, Entry};
 
 /// Legacy numeric code para nomes padrao do WebIDL.
 fn code_for_name(name: &str) -> i64 {
