+++ crates/rts-shared/src/globals/global_this/mod.rs
