+++ crates/rts-shared/src/globals/global_this/rt.rs
@@ -1,6 +1,6 @@
 //! Runtime impls de funcoes globais nao-classe (#208).
 
-use crate::namespaces::gc::handles::{alloc_entry, Entry};
+use rts_engine::heap::handles::{alloc_entry, Entry};
 
 fn str_from_parts(ptr: i64, len: i64) -> &'static str {
     if ptr == 0 || len <= 0 {
@@ -122,7 +122,7 @@ pub extern "C" fn __RTS_FN_GL_DECODE_URI_COMPONENT(ptr: i64, len: i64) -> u64 {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::namespaces::gc::handles::{with_entry, Entry};
+    use rts_engine::heap::handles::{with_entry, Entry};
 
     fn read_str(h: u64) -> String {
         with_entry(h, |e| match e {
