+++ crates/rts-shared/src/globals/json/mod.rs
