+++ crates/rts-shared/src/globals/mod.rs
@@ -13,3 +13,10 @@ pub mod weakmap;
 pub mod weakset;
 pub mod weakref;
 pub mod finalization_registry;
+pub mod regexp;
+pub mod json;
+pub mod json5;
+pub mod intl;
+pub mod dom_exception;
+pub mod global_this;
+pub mod date;
